backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
sentence-transformers
faiss-cpu
pdfplumber
numpy
python-multipart
